// microshop-ui/src/app/api/stock/model/category.ts
/**
 * MicroShop.StockAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Product } from './product';


export interface Category { 
    id?: number;
    name?: string | null;
    products?: Array<Product> | null;
}