// microshop-ui/src/app/api/basket/model/basket-item.ts
/**
 * MicroShop.BasketAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BasketItem { 
    productId?: string;
    productName?: string | null;
    price?: number;
    quantity?: number;
    pictureUrl?: string | null;
}